# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Adoption Service API
  description: API for managing pets, pet pictures, and adoption applications.
  version: 1.0.0
  contact:
    name: Adoption Service Team
    email: support@example.com
servers:
  - url: http://localhost:8080
paths:
...
  /applications:
    get:
      summary: List adoption applications
      description: >-
        Organization users see all applications. Regular users only see their own applications.
        Results are paginated; response is a Page object with content, totalElements, and totalPages.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page index
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: createdAt,desc
          description: Sort property and direction, e.g. createdAt,desc or id,asc
      responses:
        '200':
          description: Paginated list of applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPage'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Submit an adoption application
      description: The authenticated user is the applicant.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
      responses:
        '201':
          description: Created application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications/{applicationId}:
    parameters:
      - name: applicationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get an adoption application
      description: >-
        Organization users can read any application. Regular users can only read their own applications.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Partially update an adoption application
      description: >-
        Any authenticated user can patch an application. Only the fields present in the
        payload are updated; status is always required. Status transitions are constrained
        by role: REGULAR owners may move their own NEEDS_INFO application to PENDING;
        ORGANIZATION may move PENDING to NEEDS_INFO/APPROVED/REJECTED and APPROVED to
        REJECTED/COMPLETED. Marking an application COMPLETED also marks the pet ADOPTED.
        Regular users may only patch applications they own.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationPatchRequest'
      responses:
        '200':
          description: Updated application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Validation failed or application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied (regular user patching an application they do not own)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Status transition not allowed by the role policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApplicationCreateRequest:
      type: object
      required:
        - petId
        - applicantName
        - applicantEmail
        - applicantPhone
      description: The authenticated user is the applicant.
      properties:
        petId:
          type: integer
          format: int64
        applicantName:
          type: string
        applicantEmail:
          type: string
          format: email
        applicantPhone:
          type: string
        message:
          type: string
          nullable: true
    ApplicationPatchRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        applicantName:
          type: string
          nullable: true
        applicantEmail:
          type: string
          format: email
          nullable: true
        applicantPhone:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
...
    ApplicationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        petId:
          type: integer
          format: int64
        applicantName:
          type: string
        applicantEmail:
          type: string
          format: email
        applicantPhone:
          type: string
        message:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationPage:
      type: object
      description: Spring Data page wrapper returned by list endpoints.
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
          description: Current page index (zero-based)
        size:
          type: integer
          format: int32
          description: Page size
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    ApplicationStatus:
      type: string
      enum:
        - PENDING
        - NEEDS_INFO
        - APPROVED
        - REJECTED
        - COMPLETED
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string